requirements.txt
streamlit==1.31.1
openai==1.14.3
httpx<0.28
faiss-cpu==1.8.0
pypdf==4.2.0
tiktoken==0.6.0
numpy
python-dotenv==1.0.1
